Getting near the finish
Created skeleton methods for checking tag string inputs in the device
class. Still need to verify that the IP numbers inputted are acceptible
in the testIP method.

Created a CSV string creator which builds a string for the tags/values
which can then be printed to a CSV file. This string currently has both
the tags and the values, seperated by a \n. The java file output class
might not like that newline char and I might want to break that method
into a gettagsline and a getvaluesline and print them one line at a time
to the file. We will see.

I also fixed the entire front end, every dropbox now has all values and
each box can restore the inputted values if necessary.

just need to test the rest of the tags inputs and then write a string to
a csv file and the tool should be functional.

Will want to look into getting it to autogenerate a ticket, will
probably want it to do a text-dump in the mean time.