Fix
I was adding "" to the beginning of the arrays after importing them, but
was not passing the array with the "" at the beginning to the device
upon creation, so the device's arrays were not parallel.
I added an offset to the variables when passing them to the setTags
method on the device. This should be the fix.